created a schema folder and reorganized validation.ts broke out creategroupschema to one file
